requirements.txt
flask==2.3.3
requests==2.31.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
numpy==1.26.4
google-generativeai==0.5.2
faiss-cpu @ https://github.com/facebookresearch/faiss/releases/download/v1.7.4/faiss_cpu-1.7.4-cp310-cp310-manylinux2014_x86_64.whl




